chore(test): move some helpers to shared `helpers.ts`
